limit eviction
